refactor: extract shared attention reference helpers
