+++ tests/common/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) 2026, Minghua Shen.
+"""Shared helpers for FlashAttention NPU tests."""
+
+__all__ = []
