+++ tests/common/test_utils.py
@@ -1,2 +1,1 @@
 # Copyright (c) 2026, Minghua Shen.
-
