+++ tests/test_flash_attn_npu_v3_graph.py
@@ -8,7 +8,7 @@ if "Ascend950" in (torch_npu.npu.get_device_name() if torch_npu.npu.device_count
     pytest.skip("get_scheduler_metadata not on Ascend950", allow_module_level=True)
 
 from flash_attn_npu_3 import flash_attn_with_kvcache, get_scheduler_metadata
-from tests.test_flash_attn_npu_v3 import ref_flash_attention
+from tests.common.attention_ref import ref_flash_attention
 
 RTOL = 1e-2
 ATOL = 1e-2
